+++ client/config/paths.cjs
@@ -1,0 +1,17 @@
+const path = require('node:path');
+const fs = require('node:fs');
+
+// Make sure any symlinks in the project folder are resolved:
+// https://github.com/facebookincubator/create-react-app/issues/637
+const appDirectory = path.resolve(path.join(__dirname, '../..'));
+const resolveApp = (relativePath) => path.resolve(appDirectory, relativePath);
+
+const Paths = {
+  appBuild: resolveApp('dist/assets'),
+  appPublic: resolveApp('client/src/public/'),
+  appHtml: resolveApp('client/src/index.html'),
+  appIndex: resolveApp('client/src/javascript/app.tsx'),
+  appSrc: resolveApp('./'),
+};
+
+module.exports = Paths;
