+++ client/scripts/development.js
@@ -15,7 +15,7 @@ const checkRequiredFiles = require('react-dev-utils/checkRequiredFiles');
 const {choosePort, prepareUrls} = require('react-dev-utils/WebpackDevServerUtils');
 const path = require('node:path');
 const paths = require('../../shared/config/paths');
-const config = require('../config/webpack.config.dev');
+const config = require('../config/webpack.config.dev.cjs');
 
 const {proxy: floodServerProxy} = require('yargs').env('FLOOD_OPTION_').option('proxy', {
   default: 'http://127.0.0.1:3000',
