--- shared/config/paths.d.ts
@@ -1,11 +1,0 @@
-declare const PATHS: {
-  appBuild: string;
-  appDist: string;
-  appPublic: string;
-  appHtml: string;
-  appIndex: string;
-  appSrc: string;
-  dist: string;
-};
-
-export = PATHS;
